Update xlat expression parser to avoid a bounce through tmpls
the expression parser now calls the xlat_tokenize_word().  That
function returns xlat groups for strings, and not tmpls.

Update the %cast() function to always create its arguments as groups.
This helps when "foo%{User-Name}" is now parsed as an xlat group,
and not as an xlat which is pointing to a tmpl

xlat_process_arg_list() now calls a separate function to escape
and stringify its output

Add "is_argv" flag to xlat_exp_head_t.  This tells the xlat eval
code to NOT concatenate strings, as the function argument parser
will do that, with escaping.

various cleanups in xlat_tokenize_word()

Hoist constant strings after calling xlat_tokenize_input().  This
code should arguably be moved to xlat_tokenize_input().

xlat_argv tests now also print strings with the enclosing quotes.