hoist tmpls
if a tmpl is asked to parse %{...}. then the result is TMPL_TYPE_XLAT.

However, if the contents of that xlat are identically a tmpl, e.g.
User-Name, then there is no reason to keep the xlat.  Instead, we
hoist the tmpl we really want, and discard the intermediate xlat

this simplifies the internal data structures, and allows for
additional optimizations.  It also means that we do less work
at run-time in order to expand the tmpl.