Use call_env to identify Python function names
Firstly a function `<name1>_<name2>` is looked for.  If that is not
found (or there is no `name2`, then `<name1>` is looked for.

The Python module can be set with config option `mod_name1_name2` or
`mod_name1`.

The function can be overridden with config option `func_name1_name2` or
`func_name1`.

`name1` and `name2` are converted to lower case and any
non-alphanumercic characters are converted to underscores.

e.g. a module call in `recv Access-Request` will look for
`recv_accsss_request` and then `recv`.