just use fr_value_box_cast() in tmpl_to_type
as it means there are fewer corner cases in the code.

and add RDEBUG messages, so that the caller gets told when things
go wrong.

note that this function is only called from a few places:

tmpl_dcursor, which needs uint8_t

ldap maps, sql maps, attr_filter, and load-balance, which all
need strings.