replace %{1} with %regex.match(1)
like some other languages.

The unfortunate outcome is that this makes a lot of things much
more complex visually.  But it's a bit more consistent with the
rest of the xlat functions.  And, for the v4 way of "less magic
and fewer special cases".

For now, %{1} etc. is still functional

and enable %regex.match() in non-PCRE builds