complain when no 'send foo' section is found for reply 'foo'
However, this is not a fatal error.  Arguably, it should be a
compile-time error and not a run-time error.  But it's hard to
automatically correlate 'recv foo' with 'send reply-to-foo' as
there can be many different kinds of replies.

Having it a run-time error also means that the admin gets warned
on every packet, which means that they are slightly more likely
to read it and then do something about it.