better handle connected UDP sockets
and forbid reuse_port when we have connected UDP sockets, AND
the local IP is a wildcard.  The kernel will send replies to a
random socket, and not to the correct one!

In order to do "connected" UDP sockets with a local wildcard IP,
we will need to use recvfromto(), and then manually direct the
packets to the correct destination socket.

This also means that an "accept" socket will be needed, even for
UDP.  And, the "accept" socket will need to have a list of child
sockets.