the protocol encode xlat can't actually take a quoted string
it has to take an attribute reference, because it manually parses
the input as a tmpl_attr.

We arguably need to update the xlat function arguments to indicate
whether or not the argument expects to see a cursor.  And if so,
perhaps put a pointer to the cursor into a box of FR_TYPE_VOID.

which then lets functions like this either take a cursor of things
to encode, OR a string which is then parsed into a list of things
to encode.