minor fixes
add "todo" for consistent-keyed-balance.  This should be addressed
before the next release.

For non-redundant types, if the chosen home server is alive, then
return that immediately without going through the loop.  Otherwise
fall back to load-balance.

This is a change from ealier versions, and should help with better
load balancing, instead of always picking the "next" home server.

Update the documentation to match, and document keyed-load-balance