define EXT_KEY, to mirror EXT_REF
a UNION needs both a ref to the key, and also to have children.
However, everything else in the server either has a ref or children.
Plus, we don't want KEY refs to be automatically followed.

As such, we define a new extension type EXT_KEY, and then make it
largely identical to EXT_REF