reset stack properly on BEGIN/END PROTOCOL
don't push two frames onto the stack for BEGIN PROTOCOL, only
one is needed.  And BEGIN-PROTOCOL only pushed one frame.

reset the main dictionary on BEGIN, and again on END.

Add asserts that when we're done a cycle of BEGIN / END PROTOCOL,
the stack is where we expect it to be, and has the contents of
where we expect it to be.