Cache when trunk triggers are not found
So we don't hunt for them on every state change.

SQL trunks connections change from ACTIVE to FULL and back on every
query due to the limit of 1 query per connection - so it is unlikely
that triggers would be configured for those states, and caching that
they are not configured saves a lot of conf pair finding.