add error_cause to process_proxy_reply()
the realm routing code tracks what kind of error occured.  It then
passes the error_cause to process_proxy_reply().

If there's no proxy_reply, AND Post-Proxy-Type Fail sets

	&config:Response-Packet-Type := Protocol-Error

Then we create a proxy reply of Protocol-Error, and add Error-Cause
with the correct value.

Note that this failure case is where the proxy can't send the
packet anywhere, and not where we received a Protocol-Error after
proxying a request.