omit printing a union if it has only one child.
this simplifies the output and makes it easier to understand.
it's also the counterpart to adding the ALIAS which points from
the parent of the union, to the union member.

Other parts of the server will need to be updated to omit unions,
too.  But for now, this is good enough for tests.