print "::" more consistently for enums
but not for booleans, as true/false look bad with a "::" prefix.

The complicated checks for this condition are wrapped in an inline
function.  The various unlang / edit / xlat / etc. print routines
then manually add it.

If instead we add "::" in fr_value_box_print(), then there are
many cases where we _don't_ want to print it.  Rather than tweaking
all that through the code, we just instrument each part of the
printing code.