pull out function to verify parent DA
we can't always call PAIR_VERIFY from the append/prepend routines,
as the value-box hasn't always been initialized.

We could add a flag to the value-box to indicate whether or not
it has been initialized, but that's a lot of work, and is likely
to be wrong for a while.  Instead, we just ignore the problem for
now.