move "set type_size and length" to separate function
and fix its logic so that it's correct for all protocols, and for
the different kinds of VSAs we have in RADIUS.

Previously, if anyone had defined USR VSAs inside of the
Extended-Vendor-Specific-Attribute space, they would have been
encoded in the USR TLV format, and not the standard RADIUS 1/1
format.

While RFC 8044 allows for "vsa" to have any format, it also recommends
that they use the "tlv" format.

RFC 8044 doesn't recommend that the "evs" data type use the "tlv"
format, but it should.  If vendors use different formats for
attributes under the "evs" data type, then we will need to extend
the dictionary parser to allow attributes of type 'vendor' to
specify their own data type and length.