clean up unknown attributes
print union + child if the child is raw

update the struct encoder/decoder to create raw UNION members
as children of the union, instead of as children of the struct

add and use dict_attr_unknown_init(), which unifies the creation
and checking of unknown attributes.  So that we can enforce
common rules and restrictions.