hoist checks for attribute size to next_encodable
so that we don't hit assertions with unknown attributes.  The
longer term fix is to allow raw attributes under VSAs, even if
they do break the hierarchy.

Also update the encoder so that when it switches over types, the
"default" case is the error one.  This ensures that the functions
only get called with data types that they support.