manually switch to local / openssl versions of MD4 and MD5.
The fuzzer doesn't call all of the OpenSSL initialization functions,
so if it calls any MD5 routines in RADIUS, it crashes.  If we try to
call fr_openssl_fips_mode(), we need to link to libfreeradius-tls,
libfreeradius-server, libfreeradius-unlang, and then call
fr_openssl_init(), which does a ton of other things.

The solution is to manually swap the functions to their local
versions, OR the OpenSSL versions.  This is done in fr_openssl_init()
and in fr_openssl_free().  And for good measure, in then also in
fr_openssl_fips_mode().

The previous code could boot with FIPS disabled, set the MD5
pointers, and then at some later point FIPS would be disabled.
The MD5 pointers wouldn't be updated, and bad things could happen.

That case deosn't happen now, because the current code only changes
the FIPs status once at boot.  But it's better to fix the APIs to
be more correct.