allow some groups to print attributes as "flat"
For now, we add a special attribute flag "allow_flat", which
causes the attributes to be printed as a flat list.  But only when
there is only one child.  e.g. when we have:

	Vendor-Specific = { Cisco = { AVPair = 1 } }

we can now sometimes print it as

	Vendor-Specific.Cisco.AVPair = 1

for now, this is only done for attributes of type 'group', in order
to support the DER code.  We don't change the output for all
protocols, as doing so would involve a lot of changes to the tests.
The different output may also confuse the average admin, who would
sometimes see nested attributes, and sometimes flat ones, for no
apparent reason.

The parser has already been updated to take a flat definition
(as above), and create the fully nested version.