add "root" refs to FR_TYPE_ATTR
and use in DER.  The "root" ref is to the OID tree, and is largely
a work-around for the fact that the OIDs are not in their own
dictionary.

We also update the value-box code to get the correct enumv, and
then that lets the tests both drop any "OID-Tree" prefix from the
expected output, and also updates the "can't find OID" message to
refer to "OID-Tree" instead of the DER root.

Also add aliases in the DER root from "iso" to "OID-Tree.iso",
so that we the parser doesn't need to know anything about the enums
for FR_TYPE_ATTR